==> CTR_v5/03.knn_rank.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from tqdm import tqdm, trange
import matplotlib.pyplot as plt
%matplotlib inline

from os import listdir
from os.path import isfile, isdir, join

# 指定要列出所有檔案的目錄
mypath = './weblog_sorted'
# 取得所有檔案與子目錄名稱
files = listdir(mypath)
csv_list = []
# 以迴圈處理
for f in files:
  # 產生檔案的絕對路徑
  fullpath = join(mypath, f)
  # 判斷 fullpath 是檔案還是目錄
  if isfile(fullpath):
#     print("檔案：", f)
    if f[-4:]=='.csv':
        csv_list.append(f)
csv_list.sort()

with open('./item_data/dict_knn_table.pkl', 'rb') as file:
    dict_knn_table = pickle.load(file)

dict_knn_table['c3068248b86811e9b93facde48001122']

df_knn_rank = []

for idx, date in tqdm(enumerate(csv_list)):
    with open('./pid_pairs/' + 'pid_pairs_' + date[:-4] + '.pkl', 'rb') as file:
        pid_pair_date = pickle.load(file)

    df_knn_rank.append(pd.DataFrame(pid_pair_date, columns=['target', 'next']))

    def ranking(r):
        if r.target in dict_knn_table:
            ids = dict_knn_table[r.target]
#             ids, cos = zip(*lt)
            try:
                rank = ids.index(r.next) + 1
            except:
                rank = np.nan
        else:
            rank = np.nan
        return rank

    df_knn_rank[idx]['rank'] = df_knn_rank[idx].apply(ranking, axis=1)

with open('./knn_result/lt_df_knn_rank_122days.pkl', 'wb') as file:
    pickle.dump(df_knn_rank, file)

---

df_knn_rank_all = pd.concat(df_knn_rank)

df_knn_rank_all

df_knn_rank_all.dropna(inplace=True)

len(df_knn_rank_all)

4671583 / 13836074

pd.set_option('display.float_format',lambda x : '%.2f' % x)
df_knn_rank_all.describe()
