==> CTR_v5/05.resorting_table-Copy1.ipynb <==
import pandas as pd
import numpy as np
import datetime
import time
import pickle
from tqdm import tqdm
import matplotlib.pyplot as plt
%matplotlib inline
from collections import Counter
from random import choices, sample
from gensim.models import Word2Vec
import gc
import keras

PART = '2'

with open('./dict_knn_table/' + 'knn_table_' + PART + '.pkl', 'rb') as file:
    dict_table = pickle.load(file)

with open('./item_data/en_behavior_emb.pkl', 'rb') as file:
    emb = pickle.load(file)

model = keras.models.load_model('./saved_model/1434877_0424.h5')

for iid in tqdm(list(dict_table.keys())):
#     topn, cos = zip(*dict_table[iid])
    topn = dict_table[iid]
    score = [ float(model.predict(np.hstack(( emb[iid], emb[n] )).reshape(1, 256))[0][0]) for n in topn ]
    new_rank = list(zip(topn, score))
    new_rank.sort(key=lambda x : x[1], reverse=True)
    dict_table[iid] = new_rank

with open('./dict_NN_table/' + 'NN_table_' + PART + '.pkl', 'wb') as file:
    pickle.dump(dict_table, file)
