==> CTR_v5/05-1.split_dict_knn_table.ipynb <==
import pandas as pd
import numpy as np
import datetime
import time
import pickle
from tqdm import tqdm
import matplotlib.pyplot as plt
%matplotlib inline

with open('./item_data/dict_knn_table.pkl', 'rb') as file:
    d = pickle.load(file)

lt_dict_tables = []

lt_dict_tables.append(dict(list(d.items())[:100000]))
lt_dict_tables.append(dict(list(d.items())[100000:200000]))
lt_dict_tables.append(dict(list(d.items())[200000:300000]))
lt_dict_tables.append(dict(list(d.items())[300000:]))

print([len(x) for x in lt_dict_tables])

for idx, table in enumerate(lt_dict_tables):
    with open('./dict_knn_table/' + 'knn_table_' + str(idx+1) + '.pkl', 'wb') as file:
        pickle.dump(table, file)
