==> CTR_v5/13.new_switch-C.ipynb <==
import pandas as pd
import numpy as np
import datetime
import time
import pickle
from tqdm import tqdm
import matplotlib.pyplot as plt
%matplotlib inline
from collections import Counter
from random import choices, sample
from gensim.models import Word2Vec
import gc
pd.set_option('display.float_format',lambda x : '%.2f' % x)
from math import log10

with open('./NN_result/lt_df_NN_rank_122days.pkl', 'rb') as file:
    lt_df_NN_rank = pickle.load(file)

full_table = pd.concat(lt_df_NN_rank[15:15+54], ignore_index=True)

full_table['dist'] = full_table.apply(lambda df: df['rank'] - df['NN_rank'], axis=1)

full_table

full_table.describe()

table_grby = full_table.groupby(['target'])
pd.DataFrame(table_grby.size())

df_table = pd.merge(full_table, pd.DataFrame(table_grby.size()), on=['target'], how='left')

df_table.rename(columns = {0:'occur'}, inplace=True)

# df_table.drop_duplicates(inplace=True, ignore_index=True)

df_table

df_table.describe()

df_table.groupby('occur').mean().describe()

---

target_grby = table_grby # df_table.groupby(['target'])

target_array = df_table['target'].unique()
training_table = pd.DataFrame({'target': target_array})
training_table = pd.merge(training_table, pd.DataFrame(table_grby.size()), on=['target'], how='left')
training_table.rename(columns = {0:'occur'}, inplace=True)

training_table['rank'] = training_table.apply(lambda df: target_grby.get_group(df['target'])['rank'].mean(), axis=1)
training_table['NN_rank'] = training_table.apply(lambda df: target_grby.get_group(df['target'])['NN_rank'].mean(), axis=1)
training_table['label'] = training_table.apply(lambda df: df['rank'] > df['NN_rank'], axis=1)
training_table['dist'] = training_table.apply(lambda df: target_grby.get_group(df['target'])['dist'].mean(), axis=1)

training_table

training_table['label'].value_counts()

def trans(df):
    if df['label'] and df['occur']==1:
        if df['dist'] < 25:
            return False
    return df['label']

training_table['label_x'] = training_table.apply(trans , axis=1)

training_table

training_table['label_x'].value_counts()

---

df_table_testday = pd.concat(lt_df_NN_rank[15+54:], ignore_index=True)

df_table_testday.drop_duplicates(inplace=True, ignore_index=True)

df_table_testday.describe()

switch_table = training_table.drop(columns=['rank', 'NN_rank', 'dist'])

df_table_testday_switch = pd.merge(df_table_testday, switch_table, on=['target'], how='left')

from multiprocessing import  Pool
def parallelize_dataframe(df, func, n_cores=8):
    df_split = np.array_split(df, n_cores)
    pool = Pool(n_cores)
    df = pd.concat(pool.map(func, df_split))
    pool.close()
    pool.join()
    return df

def add_features(df_table_testday_switch):
    df_table_testday_switch['Ground_Truth'] = df_table_testday_switch.apply(lambda df: df['rank'] > df['NN_rank'], axis=1)
    df_table_testday_switch['switch'] = df_table_testday_switch.apply(lambda df: df['NN_rank'] if df['label'] else df['rank'], axis=1)
    df_table_testday_switch['switch_fillna_NN'] = df_table_testday_switch.apply(lambda df: df['NN_rank'] if np.isnan(df['label']) else df['switch'], axis=1)
    df_table_testday_switch['switch_fillna_knn'] = df_table_testday_switch.apply(lambda df: df['rank'] if np.isnan(df['label']) else df['switch'], axis=1)
#     df_table_testday_switch['switch_x'] = df_table_testday_switch.apply(lambda df: df['NN_rank'] if df['label_x'] else df['rank'], axis=1)
    return df_table_testday_switch

df_table_testday_switch = parallelize_dataframe(df_table_testday_switch, add_features)

df_table_testday_switch

df_table_testday_switch[df_table_testday_switch['label'].isnull()]

df_table_testday_switch.describe()

from sklearn.metrics import classification_report,confusion_matrix
print(classification_report(df_table_testday_switch['Ground_Truth'].tolist(),df_table_testday_switch['label'].fillna(True).tolist()))
